+++ satella/coding/concurrent/monitor.py
@@ -3,9 +3,9 @@ import copy
 import threading
 import typing as tp
 
-from ..decorators.decorators import wraps
+from satella.coding.decorators.decorators import wraps
 
-from ..typing import K, V, T
+from satella.coding.typing import K, V, T
 
 
 class Monitor: