+++ satella/os/__init__.py
@@ -1,7 +1,7 @@
-from .daemon import daemonize
-from .misc import suicide, is_running_as_root, whereis
-from .pidlock import PIDFileLock
-from .signals import hang_until_sig
+from satella.os.daemon import daemonize
+from satella.os.misc import suicide, is_running_as_root, whereis
+from satella.os.pidlock import PIDFileLock
+from satella.os.signals import hang_until_sig
 
 __all__ = [
     'daemonize', 'whereis',