+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # v2.23.4
+
+* added AutoflushFile
+* added correct read() to DevNullFilelikeObject
+* Proxy got _get_object()
+* added tell(), seekable(), truncate() and seek() to DevNullFilelikeObject
+